Add header for platform abstraction.
N2kDef defines the requirements from the library side. It needs millis() to get the
current uptime, and delay() to be able to delay execution. The header also comes with
a set of macros to be able to store constant data in a more efficient way. Most
platforms do not have a specific way of handling const data, but AVR8 needs explicit
handling to store constant data on flash without copying it to RAM.