ADD TYPES USED BY N2k WINDLASS NETWORK MESSAGES
A collection of enum definitions relating to PGNs

- 128776 Windlass Control Status;
- 128777 Windlass Operating Status;
- 128778 Windlass Monitoring Status.

Mostly these are peculiar to the windlass domain, but some enumss
are more generic and have not been previously defined as such in
the existing version of src/N2kTypes.h.

The language used in the definitions is explicitly that used in
the N2K specification documents with each enum defined in terms of
its N2K data definition code and its N2K data definition name.

Where appropriate (e.g. in tN2kDD002) enum member values are
overloaded to capture all of the semantics defined in the N2K
specification.